agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Agentic customer support using LangChain, Qdrant, and FastAPI"
requires-python = ">=3.11"

dependencies = [
    "fastapi[standard]<1.0.0,>=0.114.2",
    "pydantic>2.0",
    "pydantic-settings<3.0.0,>=2.2.1",
    "pyjwt<3.0.0,>=2.8.0",

    "langchain-core>=1.2.15",
    "langgraph>=1.0.9",
    "langchain-groq>=1.1.2",
    "langchain-community>=0.3.15",

    "qdrant-client[fastembed]==1.9.1",
    "redis==5.0.6",
    "httpx==0.27.0",

    "python-multipart==0.0.9",
    "asyncpg==0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
    "httpx==0.27.0",   # for FastAPI TestClient
    "ruff==0.4.9",
    "mypy==1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
    "ruff==0.4.9",
    "mypy==1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]   # pycodestyle + pyflakes + isort

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]